First pass at `jj-describe'
It works. However there are several UX issues to address. The first one
is that one needs to manually save the buffer and close in order to
yield back to jj to store the describe. We should define a mode so that
we can bind C-c C-c to record the description. The mode can be set if
the buffer ends in ".jjdescription".

🤟 This description was written from Emacs using jujutushi.el 🤟